landing page : clean code, separate widget, fix back button color, delete theme button
