+++ lib/constants/color_xetia.dart
@@ -1,6 +1,7 @@
 import 'dart:ui';
 
 const Color kOrangeXetia = const Color(0xfffbb216);
+const Color kLightOrangeXetia = const Color(0xffffc953);
 const Color kHintColor = const Color(0xffe0e1e1);
 const Color kWhite = const Color(0xffffffff);
 const Color kBlack = const Color(0xff000000);
