+++ lib/main.dart
@@ -17,6 +17,10 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return GetMaterialApp(
-        debugShowCheckedModeBanner: false, title: 'Shop_Xetia', theme: themeController.theme, home: signInController.hasLoggedIn);
+      debugShowCheckedModeBanner: false,
+      title: 'Shop_Xetia',
+      theme: themeController.theme,
+      home: signInController.hasLoggedIn,
+    );
   }
 }
