+++ lib/ui/_ui.dart
@@ -13,5 +13,3 @@ export 'shop/shop_screen.dart';
 export 'landing_page/sign_up_ui.dart';
 export 'landing_page/recovery_ui.dart';
 export 'landing_page/sign_in_ui.dart';
-export 'landing_page/sign_up_name_ui.dart';
-export 'landing_page/sign_up_password_ui.dart';
