+++ lib/ui/components/login_method/_component.dart
@@ -3,3 +3,7 @@ export 'email_login.dart';
 export 'facebook_login.dart';
 export 'gmail_login.dart';
 export 'apple_login.dart';
+
+export 'bottom_landing_page.dart';
+export 'header_landing_page.dart';
+export 'tab_landing_page.dart';
