+++ lib/ui/components/login_method/tab_landing_page.dart
@@ -1,0 +1,64 @@
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:xetia_shop/constants/_constants.dart';
+import 'package:xetia_shop/controllers/_controllers.dart';
+
+class TabNameLandingPage extends StatelessWidget {
+  final LoginController loginController = Get.find<LoginController>();
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        Spacer(),
+        Expanded(
+            flex: 3,
+            child: GestureDetector(
+                onTap: () => loginController.toggle(),
+                child: Text("Login",
+                    style: context.textTheme.headline3,
+                    textAlign: TextAlign.center))),
+        Expanded(
+            flex: 3,
+            child: GestureDetector(
+                onTap: () => loginController.toggle(),
+                child: Text("Register",
+                    style: context.textTheme.headline3,
+                    textAlign: TextAlign.center))),
+        Spacer()
+      ],
+    );
+  }
+}
+
+class TabIndicatorLandingPage extends StatelessWidget {
+  final LoginController loginController = Get.find<LoginController>();
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        Spacer(),
+        Expanded(
+            flex: 6,
+            child: Obx(() => AnimatedAlign(
+                  duration: Duration(milliseconds: 250),
+                  alignment: loginController.isLogin.value == true
+                      ? Alignment.centerLeft
+                      : Alignment.centerRight,
+                  child: Container(
+                    width: widthApp * 3 / 8,
+                    height: 10.0,
+                    decoration: BoxDecoration(
+                      borderRadius: BorderRadius.circular(20.0),
+                      color: const Color(0xfffcb216),
+                    ),
+                  ),
+                ))),
+        Spacer()
+      ],
+    );
+  }
+}
